Handle existing eBPF hooks
If a TC hook already exists when setting up the eBPF program we'll
simply go with it and use that for now. We've also added a setting
allowing for the deletion of the backing hook even if we didn't
create it.

We could also study whether we want to simply delete and recreate the
hook if it already exists, but that could be troublesome in systems
where there are other eBPF programs running...