Add a match-all marking strategy
This new strategy, controlled by the `GLOWD_FLOW_LABEL_MATCH_ALL`
constant, will match all the egressing traffic. Please bear in
mind this mechanism relies on an all-0 flowHash being inserted
in the lookup map. That is, the flowHash with which we index
the connection map will be all 0s.